requirements.txt
openai>=1.1.0
langchain>=0.0.310
langchain-openai>=0.0.2
langchain-community>=0.0.10
fastapi>=0.104.1
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.4.2
pypdf>=3.17.0
pytesseract>=0.3.10
pillow>=10.1.0
chromadb>=0.4.18
python-dotenv>=1.0.0 